Refactored multiple static utility methods out of main plugin class, into a utility class
