+++ src/main/java/com/dkabot/DkabotShop/command/HistoryCommandExecutor.java
@@ -14,6 +14,7 @@ import com.avaje.ebean.PagingList;
 import com.avaje.ebean.Query;
 import com.dkabot.DkabotShop.persistence.HistoryEntity;
 import com.dkabot.DkabotShop.DkabotShop;
+import com.dkabot.DkabotShop.util.DkbotUtils;
 
 public class HistoryCommandExecutor implements CommandExecutor {
 
@@ -60,7 +61,7 @@ public class HistoryCommandExecutor implements CommandExecutor {
                 return true;
             }
             if (material != null) {
-                if (plugin.illegalItem(material)) {
+                if (DkbotUtils.illegalItem(material)) {
                     sender.sendMessage(ChatColor.RED + "Disallowed Item!");
                     return true;
                 }
